requirements.txt
asgiref
certifi
charset-normalizer
crispy-bootstrap5
crispy-tailwind
distlib
dj-rest-auth
Django
django-allauth
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
djangorestframework_simplejwt
filelock
gunicorn
idna
Markdown
MouseInfo
numpy
opencv-contrib-python
packaging
pillow
pipenv
platformdirs
PyAutoGUI
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
razorpay
requests
setuptools
sqlparse
sslcommerz-lib
tzdata
urllib3
virtualenv
waitress
wheel
